restructure container Signed-off-by: Anna N <thuc.nguyen@st.ovgu.de>
