+++ src/utils/gcs_utils.py
@@ -4,94 +4,8 @@ import joblib
 # import json
 import io
 # import os
-from utils.config import Config
-from utils.logger import logger
-
-# def upload_dataframe_to_gcs(dataframe, bucket_name, destination_blob_name):
-#     """Uploads a pandas DataFrame to GCS as a CSV file."""
-#     client = storage.Client()
-#     bucket = client.bucket(bucket_name)
-#     blob = bucket.blob(destination_blob_name)
-
-#     csv_buffer = io.StringIO()
-#     dataframe.to_csv(csv_buffer, index=False)
-#     blob.upload_from_string(csv_buffer.getvalue(), content_type='text/csv')
-#     print(f"DataFrame uploaded to gs://{bucket_name}/{destination_blob_name}")
-
-# def download_dataframe_from_gcs(bucket_name, source_blob_name):
-#     """Downloads a CSV file from GCS and returns it as a pandas DataFrame."""
-#     client = storage.Client()
-#     bucket = client.bucket(bucket_name)
-#     blob = bucket.blob(source_blob_name)
-
-#     # Download to a string in memory
-#     csv_string = blob.download_as_text()
-#     dataframe = pd.read_csv(io.StringIO(csv_string))
-#     print(f"DataFrame downloaded from gs://{bucket_name}/{source_blob_name}")
-#     return dataframe
-
-# def upload_model_to_gcs(model, bucket_name, destination_blob_name):
-#     """Uploads a joblib-serialized model to GCS."""
-#     client = storage.Client()
-#     bucket = client.bucket(bucket_name)
-#     blob = bucket.blob(destination_blob_name)
-
-#     # Serialize model to bytes in memory
-#     # model_bytes = joblib.dump(model)
-#     blob.upload_from_string(model, content_type='application/octet-stream')
-#     print(f"Model uploaded to gs://{bucket_name}/{destination_blob_name}")
-
-# def download_model_from_gcs(bucket_name, source_blob_name):
-#     """Downloads a joblib-serialized model from GCS and deserializes it."""
-#     client = storage.Client()
-#     bucket = client.bucket(bucket_name)
-#     blob = bucket.blob(source_blob_name)
-
-#     # Download model bytes to memory
-#     model_bytes = blob.download_as_bytes()
-#     model = joblib.load(model_bytes)
-#     print(f"Model downloaded from gs://{bucket_name}/{source_blob_name}")
-#     return model
-
-# def upload_text_to_gcs(text_content, bucket_name, destination_blob_name):
-#     """Uploads a string (e.g., feature list) to GCS as a text file."""
-#     client = storage.Client()
-#     bucket = client.bucket(bucket_name)
-#     blob = bucket.blob(destination_blob_name)
-#     blob.upload_from_string(text_content, content_type='text/plain')
-#     print(f"Text content uploaded to gs://{bucket_name}/{destination_blob_name}")
-
-# def download_text_from_gcs(bucket_name, source_blob_name):
-#     """Downloads text content from GCS."""
-#     client = storage.Client()
-#     bucket = client.bucket(bucket_name)
-#     blob = bucket.blob(source_blob_name)
-#     text_content = blob.download_as_text()
-#     print(f"Text content downloaded from gs://{bucket_name}/{source_blob_name}")
-#     return text_content
-
-# def upload_series_to_gcs(series, bucket_name, destination_blob_name):
-#     """Uploads a pandas Series to GCS as a CSV file."""
-#     client = storage.Client()
-#     bucket = client.bucket(bucket_name)
-#     blob = bucket.blob(destination_blob_name)
-
-#     csv_buffer = io.StringIO()
-#     series.to_csv(csv_buffer, header=False) # Series to_csv doesn't need index=False, but header=False is good for lookup
-#     blob.upload_from_string(csv_buffer.getvalue(), content_type='text/csv')
-#     print(f"Series uploaded to gs://{bucket_name}/{destination_blob_name}")
-
-# def download_series_from_gcs(bucket_name, source_blob_name):
-#     """Downloads a CSV file from GCS and returns it as a pandas Series."""
-#     client = storage.Client()
-#     bucket = client.bucket(bucket_name)
-#     blob = bucket.blob(source_blob_name)
-
-#     csv_string = blob.download_as_text()
-#     # Read as DataFrame first then squeeze to Series, handling potential header from training
-#     series = pd.read_csv(io.StringIO(csv_string), index_col=0, header=None).squeeze("columns")
-#     print(f"Series downloaded from gs://{bucket_name}/{source_blob_name}")
-#     return series
+from src.utils.config import Config
+from src.utils.logger import logger
 
 # Helper to load model from GCS
 def load_model_from_gcs():
